Fix some x64 specific C4244 warnings.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4986 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DSUtil/DSUtil.cpp
@@ -53,7 +53,7 @@ void DumpStreamConfig(TCHAR* fn, IAMStreamConfig* pAMVSCCap)
 	f.WriteString(s);
 
 	if (size == sizeof(VIDEO_STREAM_CONFIG_CAPS)) {
-		for (ptrdiff_t i = 0; i < cnt; i++) {
+		for (int i = 0; i < cnt; i++) {
 			AM_MEDIA_TYPE* pmt = NULL;
 
 			VIDEO_STREAM_CONFIG_CAPS caps;
